Reject transfers to the same user
There is a logical error currently present in the transfer database
logic which results in no deduction for a transfer to ones self. By
making changes in parallel local variables rather than in the db, an
old version of the data read for the user balance is used. This allows
users to print money by transferring to themselves through the
webgui.

This commit doesn't fix the fundamental issue, but just exits early
for transfers to the same user. A longer-term look at this logic is
still needed, there could be other bugs and edge cases lurking.