webadmin: Don't hamburger without any menu items
The navbar has only one menu item (just the brand button to go to the
homepage) which will remain visible even when collapsed to a
hamburger. This results in a hamburger appearing on small screens with
no included menu items. This commit changes the nav config to leave it
expanded at all screen sizes. If buttons are ever added to the navbar
this should be reconsidered!