make: Use npm clean-install instead of user install
The `npm install` command is intended for interactive terminal use,
and will update the lock file inconsistently. `npm ci` is intended for
this purpose, and does a completely clean install from the latest lock
file. It will error out if a clean install is not possible, rather
than attempting a fix like `npm install`.