webadmin: Factor customer model to standalone npm package
The customer model is duplicated between the webadmin tool and the
webgui. To keep changes to the model in sync and reduce code
duplication, this commit factors the model to a standalone package
which can be included in other packages for different UI components.

Due to details with how npm handles local file dependencies, the
package directory has to have the same name and relative relationship
to packages that depend on it in both development and production. This
could be worked around at build time, but weakens the nice guarantees
provided by npm ci that the build exactly matches the dev environment.