package: Include distribution release version in package version
The debian archive format does not allow for multiple distributions to
have packages of the exact same version if the packages are not the
same. Essentially only one package can be stored in the pool at the
absolute version number. Since we want to release packages that are
the same "version", but will be compiled differently for the different
distributions, we need to include the distribution information in the
version itself.