conf: Include newline in sysctl.conf edit
The replaceAll command seems kind of strange in that it matches an
entire line, but then on edit does not write an entire line, only
inserting the specific expression *without a newline by default*. This
commit adds a newline to the inserted ip forward line in
sysctl.conf. Without this newline, a corrupted sysctl entry will be
created with a mix of whatever content was on the next line in the
file.