webgui common-models: Add plumbing for temporary test databases
This commit adds knex migrations and seed files to setup a consistent
test database from source control. This infrastructure is added to the
webgui integration tests to build temporary databases before each
test. The knexfile and knex dependency remain in the
colte-common-models package, but it might make sense to lift the knex
infrastructure to the top-level packages in the future. (This is
somewhat hard since we have multiple top-level packages that all
reference the same shared database).