common-models: Use the transaction instance form consistently
The customer model used a mix of the transaction instance form and the
transacting promise form of specifying a transaction. This
inconsistency could cause confusion, and is not a typical way to use
the underlying framework. The instance form is more explicit, and was
added by the framework authors due to how easy it is to mess up the
promise form. Rollback and commit will be called based on if there are
any errors in the underlying promise chain.