common-models: Add for update annotations to early select statements
In MySQL transactions may deadlock when a select is done early in the
transaction and the same rows are updated later in the
transaction. Since the purchase and transfer functions usually update,
we can add "FOR UPDATE" to the early select to indicate a write lock
should be acquired instead of just a read lock.