"subnet" -> "pdn" in update_smf
We previously changed "subnet" to "pdn" in smf and upf conf files due to schema changing, but one instance of "pdn" still exists in update_smf. As currently written, every time we call colteconf update it just adds a value to the subnet field instead of clearing out the old values.