webgui: Handle the case a customer is not present in the db
Instead of just crashing, render a page with unknown amounts when a
customer is not present in the database. This partially addresses #55,
although it would be good if the UI had more information in this
case. This should not cause issues with purchases or transfers, since
those routes already check if the user is from a known IP independently.