webgui: Update tests for postgres cleanup
Postgres is a little more conservative than mysql, and requires that
the connection where a drop database call is issued isn't actually
connected to the database that will be dropped. It also requires all
open connections to the database be closed before it is dropped. This
commit keeps a handle to the original connection used to create the
test database and cleans up the test database (including any dangling
open connections) after test completion.