webadmin webgui: Use absolute page paths for formatter compatibility
Prettier relies on the glimmer AST parser to format handlebars
templates, and glimmer doesn't allow for relative paths. This is
unfortunate and will hopefully be addressed with a future prettier
release, but for now we'll just use absolute paths since we rarely
used relative paths to begin with, and the current uses work just as
well with absolute paths from the root page element.