webadmin webgui common-models: Apply latest migrations on start
When the server starts, ensure the database has the latest migrations
applied. This is somewhat risky, since it means that migrations could
be applied when a user does not explicitly request for them to be
applied. In the future some kind of colte admin tool could handle
migrations instead of automatically running them.