python: Update coltedb for postgres
This commit updates the coltedb script to connect to the new postgres
database and use the up to date schema. The script currently has some
edge cases and doesn't follow best practices for error handling, and
could be greatly improved! This commit attempts to make the minimum
change possible to work with postgres while ignoring these issues.