make: Further sanitize git describe version string
When developing with additional commits since a tag, `git describe
--tags` returns a string of the form '0.15.2-5-g0c160ccd'. Debian
versions have very specific parsing rules for the "-" character. We
previously replaced the "-" separating the commit hash string with the
first sed rule, `sed s/-g/+g/g`, but ignored the "-" separating the
tag from the number of commits. This "-" also needs to be replaced to
make a valid debian version. This commit replaces it with a "~", which
is an allowed non-alphanumeric character per
https://www.debian.org/doc/debian-policy/ch-controlfields.html#s-f-version,
and still allows the commit hash to be uniquely identified by
splitting on the "+", which is also allowed.