nfpm, make: Split colte package into smaller components
As discussed in issue https://github.com/uw-ictd/colte/issues/107, we
have use cases now where networks would like to use the turnkey NITB
config provided by CoLTE, but do not necessarily need the prepaid
billing system and its associated databases and dependencies. This
commit splits the project into two different fundamental parts, the
core network and the prepaid billing suite.

For backwards compatibility it provides a high-level "colte" package
that maintains the current behavior of the unified package, installing
the billing suite and a 4G core network.

As we're starting to work with other types of core networks, including
5G cores and 4G/5G NSA cores, it also adds an abstraction where the
packages depend on a generic "colte-cn", which is provided by a
specific core network instance (to start colte-cn-4g).

The colte-essential package contains only stubs of the shared
configuration tool, database managment tool, and unified config
file. It is a dependency of both colte-cn-4g and colte-prepaid, to
allow either the standlone core or core and billing suite to be
installed without conflicting on the shared config tool files. The
config and user management tools can determine which components are
installed at runtime.

Since we are intending to use the different packages to allow
installing a subset of components, rather than using them to support
mixed versions, this commit specifies hard version dependencies
between the components for now. We will see if this causes issues in
practice, since we don't want to take on the maintenance burden of
testing mixed version compatibility at this time.