webgui: Allow higher api rate limit when testing
Some of the integration tests intentionally generate many parallel
transactions to attempt to test consistency, which would trigger the
production rate limit. For now this commit increases the rate limit to
allow for rapid testing, but probably should be re-thought and be more
principally defined per test.

Ideally we should actually design the rate-limited behavior and test
it itself!