make nfpm: Include distro release in version numbers
The provides and depends relationsips require that the distro
version (which was previously encoded into the debian-revision part of
the overall version number) be included up front. This commit appends
the distro release version directly to the package version to avoid
conditional logic in the nfpm yaml files about whether to include the
revision or not.