nfpm: Specify that split packages conflict with original colte deb
When upgrading to the new split colte packages, care must be taken
that the old monolithic colte package is removed first, before the new
split packages are installed, allowing the split packages to take
ownership of the files formerly owned by the single colte
package. Debian recommends using "replaces" and "breaks", but "breaks"
is not supported in nfpm, so this commit used the stronger "conflicts"
instead, which should be okay for this case. See
https://www.debian.org/doc/debian-policy/ch-relationships.html#s-replaces
for more information.