python: Fix exec regression
In switching to subprocess instead of os exec in commit
24af9e566ef2e3226a04ce5d58b9bff02246e927, the call to open5gs db was
errorneously only partially switched to a parameter list instead of a
single string invocation. Essentially open5gsdb was being called with
only three parameters, one of which was a single string (containing
spaces) with all of the values that should have been other parameters.