python: Remove coltedb dependency on psycopg2
psycopg2 is only needed when communicating with postgres, which should
only be required when using the metering functionality. The base
install should not require a database connection. Before this commit
the base coltedb wrapper was creating a database connection and then
passing it into the different implementation functions. This meant the
database connection library was required, even if the connection was
never needed. This commit moves the connection creation into each
prepaid billing function.

This does result in a lot of duplication, and in-general the
organization of the coltedb base script and the specialization scripts
could be improved a lot.