webgui: Render idr with fewer decimal places and "." separators
IDR is relatively inflated, and common denomenations are measured in
thousands of IDR. In the latest currencies issued by the central
bank, the ones, tens, and hundreds places of the denomination are
rendered in a small font to make the bills easier to tell
apart (particularly the 10000 and 100000). This is easier to do with
the actual currency also since the bills are a different color.

This commit adds support for rendering IDR with diminutive low-digits
when the number is round. It also uses the localization specific
separators "." for thousands and "," for decimals in Indonesia.